requirements.txt
Django==4.1.3
pymongo
djangorestframework